mlops/image_classification/src/package/runtime_requirements_tflite.txt
# SPDX-FileCopyrightText: 2025 Siemens AG
#
# SPDX-License-Identifier: MIT

Pillow==10.3.0
tflite_runtime==2.14.0
opencv_python_headless==4.9.0.80
numpy==1.24.2
